Rename files to python standard. Merge module_in_frame/ and module/.
