+++ vision/boat/detect_words.py
