--- vision/module_in_frame/requirements.txt
@@ -1,2 +1,0 @@
-numpy
-opencv-python
